% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-landmarks.R
\name{as_fishmorph_landmarks}
\alias{as_fishmorph_landmarks}
\title{Coerce to fishmorph_landmarks}
\usage{
as_fishmorph_landmarks(x, ...)
}
\arguments{
\item{x}{A \code{fishmorph_landmarks}, \code{intrait_landmarks}, matrix, array or data frame.}

\item{...}{Passed to \code{\link[=fishmorph_landmarks]{fishmorph_landmarks()}}.}
}
\value{
A \code{fishmorph_landmarks} object.
}
\description{
Coerce to fishmorph_landmarks
}
