% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{fishmorph_landmark_pairs}
\alias{fishmorph_landmark_pairs}
\title{Segment -> landmark-pair map}
\usage{
fishmorph_landmark_pairs()
}
\value{
Named list of length-2 integer vectors.
}
\description{
Segment -> landmark-pair map
}
