% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{fishmorph_segment_names}
\alias{fishmorph_segment_names}
\title{Names of the 11 FISHMORPH segments}
\usage{
fishmorph_segment_names()
}
\value{
Character vector.
}
\description{
Names of the 11 FISHMORPH segments
}
