% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-landmarks.R
\name{is_fishmorph_landmarks}
\alias{is_fishmorph_landmarks}
\title{Test whether an object is a landmark configuration}
\usage{
is_fishmorph_landmarks(x)
}
\arguments{
\item{x}{Any object. \code{TRUE} for \code{fishmorph_landmarks} or the interoperable
\code{intrait_landmarks} class.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a landmark configuration
}
