% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{load_fishmorph_phylogeny}
\alias{load_fishmorph_phylogeny}
\title{Bundled global fish phylogeny}
\usage{
load_fishmorph_phylogeny()
}
\value{
An object of class \code{"phylo"} (tip labels formatted \code{"Genus.species"}).
}
\description{
Loads the FISHMORPH phylogenetic tree bundled with the package, for use with
\code{\link[=phylo_pcoa]{phylo_pcoa()}} and the \code{"missforest_phylo"} imputation option of
\code{\link[=fishmorph_trait_space]{fishmorph_trait_space()}} and \code{\link[=impute_landmarks]{impute_landmarks()}}. This is the same
\code{FishMORPH_Phylogeny.rds} object used by intraitR.
}
\seealso{
\code{\link[=phylo_pcoa]{phylo_pcoa()}}, \code{\link[=impute_landmarks]{impute_landmarks()}}, \code{\link[=fishmorph_trait_space]{fishmorph_trait_space()}}
}
