% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfishmorph-package.R
\docType{package}
\name{Rfishmorph-package}
\alias{Rfishmorph}
\alias{Rfishmorph-package}
\title{Rfishmorph: FISHMORPH morphology, landmarks and quality control}
\description{
Tools dedicated to the FISHMORPH morphological database (Brosse et al.,
2021). The package reconstructs the 21 FISHMORPH landmarks, computes the 11
body segments and the 9 dimensionless ratios, standardizes landmark geometry
and enforces the FISHMORPH digitizing conventions, projects specimens into
the FISHMORPH functional trait space, and provides visualization and
quality-control tools. It also offers a guided workflow to add new species,
using \pkg{rfishbase} to validate taxonomy and build a global freshwater fish
list.
}
\section{Reference}{

Brosse, S., Charpin, N., Su, G., Toussaint, A., Herrera-R, G.A., Tedesco,
P.A. & Villeger, S. (2021). FISHMORPH: A global database on morphological
traits of freshwater fishes. \emph{Global Ecology and Biogeography}, 30, 2330-2336.
\doi{10.1111/geb.13395}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aureletoussaint/Rfishmorph}
  \item Report bugs at \url{https://github.com/aureletoussaint/Rfishmorph/issues}
}

}
\author{
\strong{Maintainer}: Aurele Toussaint \email{aurele.toussaint@cnrs.fr} (affiliation: CNRS)

Authors:
\itemize{
  \item Aurele Toussaint \email{aurele.toussaint@cnrs.fr} (affiliation: CNRS)
}

}
\keyword{internal}
